pyproject.toml
[tool.poetry]
name = "sensors_connectivity"
version = "1.2.1"
description = "Robonomics package to read data from sensors and publish to different output channels"
authors = [
    "Vadim Manaenko <vadim.razorq@gmail.com>",
    "Mariia Bystramovich <m.bystramovich@gmail.com>",
]
license = "BSD 3-Clause License"

repository = "https://github.com/airalab/sensors-connectivity"

readme = "README.md"

packages = [
    { include = "connectivity" },
]
exclude = [
    ".gitignore",
    ]

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
pyserial = "^3.5"
pynacl = "^1.5.0"
pyyaml = "^6.0"
requests = "^2.27.1"
sentry_sdk = "^1.1.5"
robonomics-interface = "^1.5.0"
ipfshttpclient = "0.8.0a2"
pinatapy-vourhey = "^0.1.3"
netifaces = "^0.11.0"
paho-mqtt = "^1.6.1"
prometheus-client = "^0.13.1"
py-sr25519-bindings = "^0.2.0"

[tool.poetry.scripts]
sensors_connectivity = "connectivity.main:run"
test = "tests.auto_test:main"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
